Move models into a module
