+++ studygroups/models/base.py
@@ -1,0 +1,31 @@
+from django.db import models
+from django.utils import timezone
+
+
+class SoftDeleteQuerySet(models.QuerySet):
+
+    def active(self):
+        return self.filter(deleted_at__isnull=True)
+
+    def delete(self, *args, **kwargs):
+        # Stop bulk deletes
+        self.update(deleted_at=timezone.now())
+        #TODO: check if we need to set any flags on the query set after the delete
+
+
+class LifeTimeTrackingModel(models.Model):
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+    deleted_at = models.DateTimeField(null=True, blank=True)
+
+    objects = SoftDeleteQuerySet.as_manager()
+
+    def delete(self, *args, **kwargs):
+        # Don't actually delete the object, affects django admin also
+        self.deleted_at = timezone.now()
+        self.save()
+
+    class Meta:
+        abstract = True
+
+
